add optimized lee router
This version's goal is to work well with TM. The original benchmark
violated the assumption that the write-set of a transaction should
be relatively small. The benchmark was writing to LeeThread.tempgrid,
which is always old and contains thousands of small list-objects.

In this version, I switched the implementation of the grid to a
one-dimensional list. This change improved the performance for
standard PyPy, too. However, pypy-stm crashes immediately in the
JIT -> investigate.